requirements.txt
python-dotenv
transformers==4.35.2
torch==2.1.1
comet_ml
Flask